+++ LatticeBoltzmann/OrbitCamera.h
@@ -54,9 +54,9 @@ public:
 
 
 	virtual glm::mat4 getViewMatrix();
-	virtual void processKeyboardMovement(eCameraMovementDirection direction, float deltaTime);
+	virtual void processKeyboardMovement(eCameraMovementDirection direction, double deltaTime);
 	virtual void setView(eCameraView camView);
-	virtual void processMouseScroll(float yoffset);
+	virtual void processMouseScroll(double yoffset);
 	virtual void printInfo();
 
 private:
